pyproject.toml
[project]
name = "chunksmith"
version = "0.1.0"
description = "Service for ingesting and chunking documents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.30.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.54",
    "langchain-docling>=0.2.0",
    "prometheus-client>=0.21.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg>=3.2.6",
    "psycopg-binary>=3.2.6",
    "psycopg-pool>=3.2.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "python-json-logger>=3.3.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "rq[async]>=2.3.2",
    "uvicorn[standard]>=0.34.2",
    "yoyo-migrations>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
# Tell Hatch to include the 'app' directory as the package source
packages = ["app"]

[tool.uv]

[dependency-groups]
dev = [
    "anyio>=4.9.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "reportlab>=4.4.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
# Add dev dependencies here, e.g.:
# dev-dependencies = [ "pytest" ]
